Update usage.ipynb for Colab
The unzip command is no longer needed because the source code was unpacked into the repo.

However, if the notebook is launched in Colab using the badge from the README, the repo source code will not be automatically available. A commented out cell containing a git clone command was added to help with this.

A comment was also updated to indicate that the visualized motion is saved to test.mp4.